Crate bmw_log

Macros

debug

debug_all

debug_plain

do_log

error

error_all

error_plain

fatal

fatal_all

fatal_plain

get_log_option

info

info_all

info_plain

lockr

A macro that is used to lock a rwlock in read mode and return the appropriate error if the lock is poisoned.

lockw

A macro that is used to lock a rwlock in write mode and return the appropriate error if the lock is poisoned.

log

log_all

log_init

log_plain

log_rotate

need_rotate

set_log_option

trace

trace_all

trace_plain

warn

warn_all

warn_plain

Structs

LogBuilder

LogConfig

Enums

LogConfigOption

LogConfigOptionName

LogLevel

Traits

Log