Enum bmw_log::LogConfigOption

pub enum LogConfigOption {

    Colors(bool),
    Stdout(bool),
    MaxSizeBytes(u64),
    MaxAgeMillis(u128),
    Timestamp(bool),
    Level(bool),
    LineNum(bool),
    ShowMillis(bool),
    AutoRotate(bool),
    FilePath(Option<PathBuf>),
    ShowBt(bool),
    LineNumDataMaxLen(usize),
    DeleteRotation(bool),
    FileHeader(String),

}

This enum is used to get/set log settings after Log::init is called. The only setting that cannot be set after initialization is the LogConfigOption::FilePath setting. It is read only. Trying to write to it will result in an error. The function used to get these values is Log::get_config_option and the function used to set these values is Log::set_config_option.

Variants

Colors(bool)

Whether or not to display colors for this log. The default value is true.

Stdout(bool)

Whether or not to log to standard output for this log. The default value is true.

MaxSizeBytes(u64)

The maximum size in bytes before this log needs to be rotated. The default value is 1_048_576 bytes or 1 mb.

MaxAgeMillis(u128)

The maximum time in milliseconds before this log needs to be rotated. The default value is 3_600_000 ms or 1 hour.

Timestamp(bool)

Whether or not to display the timestamp with this log. The default value is true.

Level(bool)

Whether or not to display the log level with this log. The default value is true.

LineNum(bool)

Whether or not to display the line number information with this log. The default value is true for all platforms except windows. On windows there are some potential problems with the resolve_frame function as noted in the backtrace crate documentation.

ShowMillis(bool)

Whether or not to show milliseconds with this log. The default value is true.

AutoRotate(bool)

Whether or not to auto-rotate this log. The default value is true.

FilePath(Option<PathBuf>)

The optional file path that this log writes to. The default value is None.

ShowBt(bool)

Whether or not to show backtraces with this log. Backtraces are only displayed with the LogLevel::Error and LogLevel::Fatal when this configuration is enabled. The default value is true.

LineNumDataMaxLen(usize)

The maximum length of the line number data that is logged. Since the path of the filename may be long, it must be limited. The default value is 25 characters.

DeleteRotation(bool)

Whether or not to delete the log rotation with this log. This is usually only used for testing purposes when many logs would be generated and must be deleted to save space on the test system. The default value is false.

FileHeader(String)

A header line to be displayed at the top of each file produced by this logger. The default value is an empty string which is not displayed.