Macro bmw_log::debug

macro_rules! debug {
    () => { ... };
    ($a:expr) => { ... };
    ($a:expr,$($b:tt)*) => { ... };
}

Set crate::LogLevel to crate::LogLevel::Debug or log at the crate::LogLevel::Debug log level. If no parameters are specified the log level will be set. If a single parameter is specified, that string will be logged. If two or more parameters are specified, the first parameter is a format string, the additional parameters will be formatted based on the format string. Logging is done by the global logger which can be configured using the crate::log_init macro. If crate::log_init is not called, the default values are used.

Examples

 use bmw_err::Error;
 use bmw_log::*;

 debug!();

 fn main() -> Result<(), Error> {
     debug!("v1={},v2={}", 123, "def")?;

     Ok(())
 }

Note: log level must be set before using the logger or a compilation error will occur. Log level can be changed at any time and the inner most scope is used. The suggested method of use is to set the level at the top of each file and adjust as development is completed. You may start with debug or trace and eventually end up at info or warn.