Macro bmw_log::get_log_option

macro_rules! get_log_option {
    ($option:expr) => { ... };
}

Get the current value of the specified log option for the global log. The single parameter must be of the type ConfigOptionName. The macro returns a ConfigOption on success and Error on error. See crate::Log::get_config_option which is the underlying function call for full details.

Examples

 use bmw_err::Error;
 use bmw_log::*;
 use bmw_conf::*;

 debug!();

 fn test() -> Result<(), Error> {

     // first init the logger
     log_init!(DisplayColors(true))?;
     // get the configured value for FilePath
     let v = get_log_option!(ConfigOptionName::LogFilePath)?;
     // It should be the default value which is None
     assert_eq!(v, ConfigOption::LogFilePath(None));
     info!("file_path={:?}", v)?;

     Ok(())
 }