Macro bmw_log::logger

macro_rules! logger {
    ($($config:tt)*) => { ... };
}

This macro builds a crate::Log implementation and returns it. Specifically, it return a Box<dyn Log + Send + Sync>. This example below shows all of the allowed configurations that may be specified. All of these are optional.

Examples

 use bmw_err::*;
 use bmw_log::*;
 use bmw_test::*;
 use std::path::PathBuf;

 fn main() -> Result<(), Error> {
     let test_info = test_info!()?;
     let mut buf = PathBuf::new();
     buf.push(test_info.directory());
     buf.push("test.log");
     let buf = buf.display().to_string();

     // these are all the legal configurations to use.
     // most of these are the defaults. The only exceptions are:
     // MaxSizeBytes - u64::MAX (i.e. no rotation)
     // MaxAgeMillis - u128::MAX (i.e. no rotation)
     // LogFilePath - None (by default only stdout logging occurs)
     // FileHeader - "" (no header)
     let mut logger = logger!(
         MaxSizeBytes(1_024 * 1_024), // set rotation at 1mb
         MaxAgeMillis(60 * 60 * 1_000), // set rotation at 1hr
         DisplayColors(true), // whether or not to display colors on stdout
         DisplayStdout(true), // whether or not to display on stdout
         DisplayTimestamp(true), // whether or not to display the timestamp
         DisplayLogLevel(true), // whether or not to display the log level
         DisplayLineNum(true), // whether or not to display the code line number
         DisplayMillis(true), // whether or not to display millisecond precision
         LogFilePath(&buf), // path to the log file or None if no file logging
         AutoRotate(true), // whether or not to automatically rotate the log file
         DisplayBacktrace(false), // whether or not to display a backtrace on error/fatal
         LineNumDataMaxLen(30), // maximum length of line num data
         DeleteRotation(false), // whether or not to delete the rotated log file (test only)
         FileHeader("my_header"), // header to place at the top of each file
     )?;

     logger.init()?;
     logger.set_log_level(LogLevel::Debug);

     logger.log(LogLevel::Debug, "This is a test!")?;
     logger.log(LogLevel::Trace, "This will not show up!")?;

     Ok(())
 }

Invalid values

The value for MaxAgeMillis must be at least 1_000 (1 second).

The value for MaxSizeBytes must be at least 50 bytes.

The value for LineNumDataMaxLen must be at least 10 bytes.

The parent directory of LogFilePath must exist.